requirements.txt
alembic==1.6.5
aniso8601==9.0.1
click==8.0.1
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-Moment==0.11.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
pytz==2021.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
